requirements.txt
Flask
boto3
python-dotenv
cryptography